control_node/radio: clean configuration
Configuration is now passed as a structure.

Update is done in two steps, one stop in an event, and doing the actual
configuration in another event.

This way, when updating the configuration, there may not be any remaining
'to handle measure'. It's the same way as what is done in cn_consumption.