iotlab_time: ensure set_time coherency when repeated
When setting time periodically, events may occur and not have been processed
before a set_time is done.
However, they should be timed as if they where treated before set_time.

It should be ensured that events order is coherent between two set_time even if
one event before set_time is converted after set_time.

To ensure this if event is before or equal to "last_set_time", it will be timed
with previous configuration. 'or equal' is used as a measure may already have
been timed with this timestamp.